+++ src/core/agent_space_interaction.jl
@@ -213,6 +213,7 @@ function add_agent!(model::ABM{A, S}, properties...; kwargs...) where {A, S<:Abs
 end
 
 """
+<<<<<<< HEAD
 Returns the next largest agent id.
 ## Example
 ```julia
@@ -226,6 +227,10 @@ end
 model = ABM(Agent, GraphSpace(complete_digraph(5)))
 id = nextid(model) # returns 1
 ```
+=======
+    nextid(model::ABM) → id
+Return a valid `id` for creating a new agent with it.
+>>>>>>> upstream/master
 """
 nextid(model::ABM) = isempty(model.agents) ? 1 : maximum(keys(model.agents)) + 1